+++ modules/automation-account/variables-tags.tf
@@ -1,0 +1,17 @@
+variable "default_tags_enabled" {
+  description = "Option to enable or disable default tags."
+  type        = bool
+  default     = true
+}
+
+variable "automation_account_extra_tags" {
+  description = "Extra tags to add to automation account"
+  type        = map(string)
+  default     = {}
+}
+
+variable "extra_tags" {
+  description = "Extra tags to add"
+  type        = map(string)
+  default     = {}
+}