+++ modules/vm-monitoring/versions.tf
@@ -2,7 +2,17 @@
 terraform {
   required_version = ">= 0.13"
   required_providers {
-    azurerm = ">= 1.40.0"
-    null    = ">= 3"
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = ">= 1.40"
+    }
+    null = {
+      source  = "hashicorp/null"
+      version = ">= 3"
+    }
+    azurecaf = {
+      source  = "aztfmod/azurecaf"
+      version = "~> 1.1"
+    }
   }
 }